Create bias_map_fb_fe_common_cartopy.py
Created a script to use cartopy to make the bias maps instead of basemap. It is functional and looks okay, but needs aesthetic adjustment.

There were lots of little version errors that required strange fixes. such as using "np.reshape(x,-1)" to plot the x coordinate in the scatter function. It would break with a tuple otherwise